+++ src/hyperlight_host/src/sandbox/hypervisor.rs
@@ -19,6 +19,7 @@ use std::sync::OnceLock;
 
 #[cfg(mshv)]
 use crate::hypervisor::hyperv_linux;
+use crate::hypervisor::hyperv_windows;
 #[cfg(kvm)]
 use crate::hypervisor::kvm;
 
@@ -51,9 +52,7 @@ pub fn get_available_hypervisor() -> &'static Option<HypervisorType> {
                     None
                 }
             } else if #[cfg(target_os = "windows")] {
-                use crate::sandbox::windows_hypervisor_platform;
-
-                if windows_hypervisor_platform::is_hypervisor_present() {
+                if hyperv_windows::is_hypervisor_present() {
                     Some(HypervisorType::Whp)
                 } else {
                     None
